pyproject.toml
[project]
name = "comparing-wavelets-for-image-compression"
version = "0.1.0"
description = "A notebook blog-post for blog.teaochadesign.com"
authors = [
    { name = "Matthew Bilton", email = "matt@teaochadesign.com" }
]
requires-python = ">=3.8"
dependencies = [
    "beautifulsoup4>=4.10",
    "cycler>=0.11.0",
    "jupyterlab>=3.4",
    "matplotlib>=3.4",
    "matplotlib-inline>=0.1",
    "nbconvert>=6.5.0",
    "numpy>=1.19",
    "Pillow>=8.4.0",
    "PyWavelets>=1.3",
    "requests>=2.26.0",
    "scipy>=1.7",
]

[project.optional-dependencies]
test = [
    "pytest-mock>=3.6",
    "pytest-cov>=3.0",
]
dev = [
    "autopep8>=1.5",
]

[tool.autopep8]
max_line_length = 100
ignore = "E501,W6"  # or ["E501", "W6"]
in-place = true
recursive = true
aggressive = 3

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"